Move toOurThread(s) into notmuch.go
Move the conversion helper functions into notmuch.go as they only
convert notmuch.Thread objects into pinbox.Thread objects.
